#29 Import From Device Error Message
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@1521 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/org.wcs.smart.cybertracker/src/org/wcs/smart/cybertracker/util/PdaUtil.java
@@ -54,6 +54,8 @@ public class PdaUtil {
 	}	
 
 	public static void deleteTempDirectory(File tempDir) {
+		if (tempDir == null)
+			return;
 		try {
 			FileUtils.deleteDirectory(tempDir);
 		} catch (IOException e) {
